% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoLinker-Package.R
\docType{package}
\name{annoLinker-package}
\alias{annoLinker-package}
\title{Annotating genomic regions through chromatin interaction links}
\description{
Fast annotation of genomic peaks using DNA interaction data by building
interaction networks with igraph. Peaks overlapping any node in a connected
subgraph are annotated with all genes in that subgraph.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jianhong/annoLinker}
  \item Report bugs at \url{https://github.com/jianhong/annoLinker/issues}
}

}
\author{
Jianhong Ou

Maintainer: Jianhong Ou \href{mailto:jianhong.ou@morgridge.org}{jianhong.ou@morgridge.org}
}
\keyword{package}
