% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEvidence.R
\name{plotEvidence}
\alias{plotEvidence}
\title{Plot interaction network for visualization}
\usage{
plotEvidence(
  g,
  evidence,
  cluster_method = c("components", "louvain", "walktrap", "infomap"),
  ...,
  output = c("graph", "htmlWidget", "trackPlot"),
  txdb,
  org
)
}
\arguments{
\item{g}{Interaction graph output by \link{annoLinker}}

\item{evidence}{The evidence to highlighted}

\item{cluster_method}{Character, clustering method: "components" (connected
components), "louvain", "walktrap", or "infomap"}

\item{...}{parameters for cluster. see \link[igraph]{cluster_louvain},
\link[igraph]{cluster_walktrap}, and \link[igraph]{cluster_infomap}.}

\item{output}{Output of the plot.}

\item{txdb, org}{The TxDb and OrgDb object used for annotation plot.}
}
\description{
Plot interaction network for visualization
}
