% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEvidence.R
\name{plotEvidence}
\alias{plotEvidence}
\title{Plot interaction network for visualization}
\usage{
plotEvidence(
  g,
  evidence,
  cluster_method = c("components", "louvain", "walktrap", "infomap"),
  ...,
  output = c("graph", "htmlWidget", "trackPlot"),
  txdb,
  org
)
}
\arguments{
\item{g}{Interaction graph output by \link{annoLinker}}

\item{evidence}{The evidence to highlighted}

\item{cluster_method}{Character, clustering method: "components" (connected
components), "louvain", "walktrap", or "infomap"}

\item{...}{parameters for cluster. see \link[igraph]{cluster_louvain},
\link[igraph]{cluster_walktrap}, and \link[igraph]{cluster_infomap}.}

\item{output}{Output of the plot.}

\item{txdb, org}{The TxDb and OrgDb object used for annotation plot.}
}
\description{
Plot interaction network for visualization
}
\examples{
anno <- readRDS(system.file('extdata', 'sample_res.rds', package='annoLinker'))
library(org.Dr.eg.db)
library(TxDb.Drerio.UCSC.danRer10.refGene)
plotEvidence(anno$interaction_graph,
 evidence=anno$annotated_peaks$evidences[1],
 output='htmlWidget')
plotEvidence(anno$interaction_graph,
 evidence=anno$annotated_peaks$evidences[1],
 output='trackPlot',
 txdb=TxDb.Drerio.UCSC.danRer10.refGene,
 org=org.Dr.eg.db)
}
